Break ExtensionsTests.cs into one file per type tested
This change is essentially just style, no tests have been altered, added, or removed.

The tests that were about a particular extension have moved to a class named
for the type being tested (with the X509 prefix and Extension suffix omitted).
Now that the type tested is inferred from the name, the method names were
adjusted accordingly. e.g. ExtensionsTests.KeyUsageExtension_CrlSign is now
KeyUsageTests.EncodeDecode_CrlSign

The one test that wasn't about a single extension is now in the file
ExtensionsTests\ComprehensiveTests.cs, as it checks all of the extensions,
and the order of the extensions, from an existing certificate.
