+++ src/libraries/System.Security.Cryptography.X509Certificates/tests/System.Security.Cryptography.X509Certificates.Tests.csproj
@@ -34,7 +34,11 @@
     <Compile Include="ContentTypeTests.cs" />
     <Compile Include="CtorTests.cs" />
     <Compile Include="ExportTests.cs" />
-    <Compile Include="ExtensionsTests.cs" />
+    <Compile Include="ExtensionsTests\BasicConstraintsTests.cs" />
+    <Compile Include="ExtensionsTests\ComprehensiveTests.cs" />
+    <Compile Include="ExtensionsTests\EnhancedKeyUsageTests.cs" />
+    <Compile Include="ExtensionsTests\KeyUsageTests.cs" />
+    <Compile Include="ExtensionsTests\SubjectKeyIdentifierTests.cs" />
     <Compile Include="FindTests.cs" />
     <Compile Include="LoadFromFileTests.cs" />
     <Compile Include="NameTests.cs" />
